+++ examples/simpsons_quotes.py
@@ -38,4 +38,6 @@ QUOTES: Final[list[str]] = [
     "Now with vitamin R!",
     "Me fail English? Unpossible!",
     "The goggles! They do nothing!",
+    "If you don't like your job, you don't go on strike; you go in every day and do it really half assed",
+    "I don't want to look like a weirdo, I'll just wear a muumuu.",
 ]